// src/types.ts
export type ProductCategory = 
  | 'heating'
  | 'ventilation'
  | 'drinking'
  | 'cooling'
  | 'inlet'
  | 'feeding'
  | 'cage'
  | 'silo'
  | 'other';
// ...
export interface ProductSpec {
  label: string;
  value: string;
}

export interface Product {
  isFeatured?: boolean;
  id: string;
  code: string;
  name: string;
  nameEn?: string;
  category: ProductCategory;
  categoryTitle: string;
  isIndustrialMachine?: boolean;
  shortDescription: string;
  fullDescription: string;
  image: string;
  gallery?: string[];
  advantages: string[];
  applications: string[];
  specs: ProductSpec[];
  models?: string[];
  catalogPdfName?: string;
  relatedProjectIds?: string[];
  relatedArticleIds?: string[];
  faqs?: { q: string; a: string }[];
}